Add resources/prompts dispatch with ACL and split monolithic files
The proxy only handled tools/list and tools/call. The ACL schema from #55
already had resources and prompts fields on grants but they were dead code
with no runtime effect.

Wire the four missing MCP methods (resources/list, resources/read,
prompts/list, prompts/get) through the proxy using the same aliasing,
discovery, connection, and error handling patterns as tools. ACL evaluation
filters listings and gates read/get calls, with resources and prompts
always treated as read-access in v1. The CLI acl check gains --resource
and --prompt flags for operator validation.

Also split serve.rs, acl.rs and main.rs into focused module directories
for better navigability — no behavior change on the structural side.

Closes #58
