+++ src/cli/mod.rs
@@ -1,0 +1,7 @@
+mod acl;
+mod logs;
+mod server;
+
+pub use acl::handle_acl_command;
+pub use logs::handle_logs_command;
+pub use server::handle_server_command;
