+++ src/server_auth/acl/eval.rs
@@ -1,289 +1,17 @@
-use std::collections::HashMap;
-use std::fmt;
+use crate::classifier::Kind;
 
-use serde::Deserialize;
-
-use super::AuthIdentity;
-use crate::classifier::{Kind, Source, ToolClassification};
-
-// ---------------------------------------------------------------------------
-// Shared types
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Deserialize, Clone, PartialEq)]
-#[serde(rename_all = "lowercase")]
-pub enum AclPolicy {
-    Allow,
-    Deny,
-}
-
-fn default_policy() -> AclPolicy {
-    AclPolicy::Allow
-}
-
-// ---------------------------------------------------------------------------
-// Legacy schema (first-match-wins, flat rules list)
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct AclRule {
-    #[serde(default)]
-    pub subjects: Vec<String>,
-    #[serde(default)]
-    pub roles: Vec<String>,
-    pub tools: Vec<String>,
-    pub policy: AclPolicy,
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct LegacyAclConfig {
-    #[serde(default = "default_policy")]
-    pub default: AclPolicy,
-    #[serde(default)]
-    pub rules: Vec<AclRule>,
-}
-
-// ---------------------------------------------------------------------------
-// New role-based schema
-// ---------------------------------------------------------------------------
-
-/// A single server string or array of server globs.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(untagged)]
-pub enum ServerPattern {
-    Single(String),
-    Multiple(Vec<String>),
-}
-
-/// The access level for a grant.
-#[derive(Debug, Clone, PartialEq, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum AccessLevel {
-    Read,
-    Write,
-    #[serde(rename = "*")]
-    All,
-}
-
-impl AccessLevel {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            AccessLevel::Read => "read",
-            AccessLevel::Write => "write",
-            AccessLevel::All => "*",
-        }
-    }
-}
-
-/// A single grant entry within a role definition or a subject's `extra` list.
-#[derive(Debug, Clone, Deserialize)]
-pub struct Grant {
-    pub server: ServerPattern,
-    pub access: AccessLevel,
-    #[serde(default)]
-    pub tools: Vec<String>,
-    #[serde(default)]
-    pub resources: Vec<String>,
-    #[serde(default)]
-    pub prompts: Vec<String>,
-    #[serde(default)]
-    pub deny: bool,
-}
-
-/// Per-subject configuration: assigned roles + optional extra grants.
-#[derive(Debug, Clone, Deserialize)]
-pub struct SubjectConfig {
-    #[serde(default)]
-    pub roles: Vec<String>,
-    #[serde(default)]
-    pub extra: Vec<Grant>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-pub struct RoleBasedAclConfig {
-    #[serde(default = "default_policy")]
-    pub default: AclPolicy,
-    #[serde(default, rename = "strictClassification")]
-    pub strict_classification: bool,
-    #[serde(default)]
-    pub roles: HashMap<String, Vec<Grant>>,
-    #[serde(default)]
-    pub subjects: HashMap<String, SubjectConfig>,
-}
-
-// ---------------------------------------------------------------------------
-// Unified AclConfig enum with custom deserializer
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Clone)]
-pub enum AclConfig {
-    Legacy(LegacyAclConfig),
-    RoleBased(RoleBasedAclConfig),
-}
-
-impl<'de> Deserialize<'de> for AclConfig {
-    fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        let raw: serde_json::Value = serde_json::Value::deserialize(deserializer)?;
-        let obj = raw
-            .as_object()
-            .ok_or_else(|| serde::de::Error::custom("ACL config must be a JSON object"))?;
-
-        let has_rules = obj.contains_key("rules");
-        let has_roles = obj.contains_key("roles");
-        let has_subjects = obj.contains_key("subjects");
-        let is_new = has_roles || has_subjects;
-
-        if has_rules && is_new {
-            return Err(serde::de::Error::custom(
-                "ACL config cannot have both 'rules' (legacy) and 'roles'/'subjects' (new schema)",
-            ));
-        }
-
-        if is_new {
-            let rbac: RoleBasedAclConfig =
-                serde_json::from_value(raw).map_err(serde::de::Error::custom)?;
-            // Validate that all roles referenced by subjects exist in the roles map.
-            for (subject, config) in &rbac.subjects {
-                for role in &config.roles {
-                    if !rbac.roles.contains_key(role) {
-                        return Err(serde::de::Error::custom(format!(
-                            "subject '{}' references unknown role '{}'",
-                            subject, role
-                        )));
-                    }
-                }
-            }
-            Ok(AclConfig::RoleBased(rbac))
-        } else {
-            let legacy: LegacyAclConfig =
-                serde_json::from_value(raw).map_err(serde::de::Error::custom)?;
-            Ok(AclConfig::Legacy(legacy))
-        }
-    }
-}
-
-#[cfg(test)]
-impl AclConfig {
-    pub fn legacy(default: AclPolicy, rules: Vec<AclRule>) -> Self {
-        AclConfig::Legacy(LegacyAclConfig { default, rules })
-    }
-}
-
-// ---------------------------------------------------------------------------
-// ToolContext — coordinates for role-based evaluation
-// ---------------------------------------------------------------------------
-
-pub struct ToolContext<'a> {
-    pub server_alias: &'a str,
-    pub tool_name: &'a str,
-    pub classification: Option<&'a ToolClassification>,
-}
-
-pub struct ResourceContext<'a> {
-    pub server_alias: &'a str,
-    pub resource_uri: &'a str,
-}
-
-pub struct PromptContext<'a> {
-    pub server_alias: &'a str,
-    pub prompt_name: &'a str,
-}
-
-// ---------------------------------------------------------------------------
-// Decision — structured result from ACL evaluation
-// ---------------------------------------------------------------------------
-
-/// Stable identifier for the rule that determined the access decision.
-#[derive(Debug, Clone, PartialEq)]
-pub enum MatchedRule {
-    /// Legacy schema: rule at index N matched.
-    Legacy(usize),
-    /// Legacy schema default policy.
-    LegacyDefault,
-    /// RBAC: grant from role at grant-index within that role.
-    RoleGrant { role: String, index: usize },
-    /// RBAC: extra grant on a subject.
-    SubjectExtra { subject: String, index: usize },
-    /// RBAC default policy.
-    RbacDefault,
-    /// No ACL configured.
-    NoAcl,
-}
-
-impl fmt::Display for MatchedRule {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            MatchedRule::Legacy(i) => write!(f, "legacy[{i}]"),
-            MatchedRule::LegacyDefault => write!(f, "legacy:default"),
-            MatchedRule::RoleGrant { role, index } => write!(f, "{role}[{index}]"),
-            MatchedRule::SubjectExtra { subject, index } => {
-                write!(f, "{subject}.extra[{index}]")
-            }
-            MatchedRule::RbacDefault => write!(f, "default"),
-            MatchedRule::NoAcl => write!(f, "no-acl"),
-        }
-    }
-}
-
-/// Structured result from ACL evaluation, carrying the decision and its provenance.
-#[derive(Debug, Clone)]
-pub struct Decision {
-    pub allowed: bool,
-    pub matched_rule: MatchedRule,
-    pub classification_kind: Option<Kind>,
-    pub classification_source: Option<Source>,
-    pub classification_confidence: Option<f32>,
-    pub access_evaluated: Option<AccessLevel>,
-}
-
-impl Decision {
-    fn from_ctx(allowed: bool, matched_rule: MatchedRule, ctx: Option<&ToolContext>) -> Self {
-        let (kind, source, confidence) = match ctx.and_then(|c| c.classification) {
-            Some(cls) => (Some(cls.kind), Some(cls.source), Some(cls.confidence)),
-            None => (None, None, None),
-        };
-        Self {
-            allowed,
-            matched_rule,
-            classification_kind: kind,
-            classification_source: source,
-            classification_confidence: confidence,
-            access_evaluated: None,
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Unified dispatcher
-// ---------------------------------------------------------------------------
-
-/// Check if a tool is allowed for the given identity.
-/// Legacy schema uses first-match-wins; role-based uses union evaluation.
-/// Returns a structured `Decision` with provenance information.
-pub fn is_tool_allowed(
-    identity: &AuthIdentity,
-    tool_name: &str,
-    acl: &AclConfig,
-    ctx: Option<&ToolContext>,
-) -> Decision {
-    match acl {
-        AclConfig::Legacy(legacy) => legacy_is_tool_allowed(identity, tool_name, legacy, ctx),
-        AclConfig::RoleBased(rbac) => match ctx {
-            Some(c) => is_tool_allowed_rbac(identity, c, rbac),
-            None => Decision::from_ctx(
-                rbac.default == AclPolicy::Allow,
-                MatchedRule::RbacDefault,
-                None,
-            ),
-        },
-    }
-}
+use super::glob::glob_match;
+use super::types::{
+    AccessLevel, AclPolicy, AclRule, Decision, Grant, LegacyAclConfig, MatchedRule, PromptContext,
+    ResourceContext, RoleBasedAclConfig, ServerPattern, TaggedGrant, ToolContext,
+};
+use crate::server_auth::AuthIdentity;
 
 // ---------------------------------------------------------------------------
 // Legacy evaluator (first-match-wins, unchanged logic)
 // ---------------------------------------------------------------------------
 
-fn legacy_is_tool_allowed(
+pub(super) fn legacy_is_tool_allowed(
     identity: &AuthIdentity,
     tool_name: &str,
     acl: &LegacyAclConfig,
@@ -305,7 +33,7 @@ fn legacy_is_tool_allowed(
     )
 }
 
-fn matches_identity(identity: &AuthIdentity, rule: &AclRule) -> bool {
+pub(super) fn matches_identity(identity: &AuthIdentity, rule: &AclRule) -> bool {
     let subject_match = rule.subjects.is_empty()
         || rule
             .subjects
@@ -321,7 +49,7 @@ fn matches_identity(identity: &AuthIdentity, rule: &AclRule) -> bool {
     subject_match && role_match
 }
 
-fn matches_tool(tool_name: &str, patterns: &[String]) -> bool {
+pub(super) fn matches_tool(tool_name: &str, patterns: &[String]) -> bool {
     patterns
         .iter()
         .any(|pattern| glob_match(pattern, tool_name))
@@ -331,13 +59,7 @@ fn matches_tool(tool_name: &str, patterns: &[String]) -> bool {
 // Role-based evaluator (union semantics)
 // ---------------------------------------------------------------------------
 
-/// A grant tagged with its provenance (which role/subject it came from).
-struct TaggedGrant<'a> {
-    grant: &'a Grant,
-    origin: MatchedRule,
-}
-
-fn is_tool_allowed_rbac(
+pub(super) fn is_tool_allowed_rbac(
     identity: &AuthIdentity,
     ctx: &ToolContext,
     acl: &RoleBasedAclConfig,
@@ -427,7 +149,7 @@ fn is_tool_allowed_rbac(
 }
 
 /// Resolve roles and extra grants for an identity.
-fn resolve_subject<'a>(
+pub(super) fn resolve_subject<'a>(
     identity: &AuthIdentity,
     acl: &'a RoleBasedAclConfig,
 ) -> (Vec<String>, Vec<&'a Grant>) {
@@ -446,14 +168,14 @@ fn resolve_subject<'a>(
     (role_names, extra)
 }
 
-fn matches_server(grant: &Grant, server: &str) -> bool {
+pub(super) fn matches_server(grant: &Grant, server: &str) -> bool {
     match &grant.server {
         ServerPattern::Single(s) => glob_match(s, server),
         ServerPattern::Multiple(list) => list.iter().any(|s| glob_match(s, server)),
     }
 }
 
-fn matches_tool_grant(grant: &Grant, tool_name: &str) -> bool {
+pub(super) fn matches_tool_grant(grant: &Grant, tool_name: &str) -> bool {
     if grant.tools.is_empty() {
         return true;
     }
@@ -463,7 +185,7 @@ fn matches_tool_grant(grant: &Grant, tool_name: &str) -> bool {
         .any(|pattern| glob_match(pattern, tool_name))
 }
 
-fn matches_resource_grant(grant: &Grant, resource_uri: &str) -> bool {
+pub(super) fn matches_resource_grant(grant: &Grant, resource_uri: &str) -> bool {
     if grant.resources.is_empty() {
         return true;
     }
@@ -473,7 +195,7 @@ fn matches_resource_grant(grant: &Grant, resource_uri: &str) -> bool {
         .any(|pattern| glob_match(pattern, resource_uri))
 }
 
-fn matches_prompt_grant(grant: &Grant, prompt_name: &str) -> bool {
+pub(super) fn matches_prompt_grant(grant: &Grant, prompt_name: &str) -> bool {
     if grant.prompts.is_empty() {
         return true;
     }
@@ -484,21 +206,21 @@ fn matches_prompt_grant(grant: &Grant, prompt_name: &str) -> bool {
 }
 
 /// Check if a grant's access level covers read (for resources/prompts).
-fn grant_covers_read(grant: &Grant) -> bool {
+pub(super) fn grant_covers_read(grant: &Grant) -> bool {
     matches!(grant.access, AccessLevel::All | AccessLevel::Read)
 }
 
 /// Check if a tool is allowed for the given identity.
 /// Legacy schema uses first-match-wins; role-based uses union evaluation.
 /// Returns a structured `Decision` with provenance information.
-pub fn is_resource_allowed(
+pub(super) fn is_resource_allowed(
     identity: &AuthIdentity,
     _resource_uri: &str,
-    acl: &AclConfig,
+    acl: &super::AclConfig,
     ctx: Option<&ResourceContext>,
 ) -> Decision {
     match acl {
-        AclConfig::Legacy(legacy) => {
+        super::AclConfig::Legacy(legacy) => {
             // Legacy: deny read unless default is allow.
             Decision {
                 allowed: legacy.default == AclPolicy::Allow,
@@ -509,7 +231,7 @@ pub fn is_resource_allowed(
                 access_evaluated: Some(AccessLevel::Read),
             }
         }
-        AclConfig::RoleBased(rbac) => match ctx {
+        super::AclConfig::RoleBased(rbac) => match ctx {
             Some(c) => is_resource_allowed_rbac(identity, c, rbac),
             None => Decision {
                 allowed: rbac.default == AclPolicy::Allow,
@@ -523,14 +245,14 @@ pub fn is_resource_allowed(
     }
 }
 
-pub fn is_prompt_allowed(
+pub(super) fn is_prompt_allowed(
     identity: &AuthIdentity,
     _prompt_name: &str,
-    acl: &AclConfig,
+    acl: &super::AclConfig,
     ctx: Option<&PromptContext>,
 ) -> Decision {
     match acl {
-        AclConfig::Legacy(legacy) => {
+        super::AclConfig::Legacy(legacy) => {
             // Legacy: deny get unless default is allow.
             Decision {
                 allowed: legacy.default == AclPolicy::Allow,
@@ -541,7 +263,7 @@ pub fn is_prompt_allowed(
                 access_evaluated: Some(AccessLevel::Read),
             }
         }
-        AclConfig::RoleBased(rbac) => match ctx {
+        super::AclConfig::RoleBased(rbac) => match ctx {
             Some(c) => is_prompt_allowed_rbac(identity, c, rbac),
             None => Decision {
                 allowed: rbac.default == AclPolicy::Allow,
@@ -711,7 +433,7 @@ fn is_prompt_allowed_rbac(
 }
 
 /// Check if a grant's access level covers the tool's classified kind (for allow grants).
-fn grant_covers_access(grant: &Grant, kind: Option<Kind>, strict: bool) -> bool {
+pub(super) fn grant_covers_access(grant: &Grant, kind: Option<Kind>, strict: bool) -> bool {
     if grant.deny {
         return false;
     }
@@ -728,7 +450,7 @@ fn grant_covers_access(grant: &Grant, kind: Option<Kind>, strict: bool) -> bool 
 
 /// Check if a deny grant's access level covers the tool's classified kind.
 /// Same expansion logic as allow, but ignores the `deny` flag guard.
-fn deny_covers_access(grant: &Grant, kind: Option<Kind>, strict: bool) -> bool {
+pub(super) fn deny_covers_access(grant: &Grant, kind: Option<Kind>, strict: bool) -> bool {
     if strict && matches!(kind, Some(Kind::Ambiguous)) {
         return true; // strict mode: deny always catches ambiguous
     }
@@ -739,102 +461,12 @@ fn deny_covers_access(grant: &Grant, kind: Option<Kind>, strict: bool) -> bool {
     }
 }
 
-// ---------------------------------------------------------------------------
-// Glob matching (shared by both schemas)
-// ---------------------------------------------------------------------------
-
-/// Glob matching: supports `*` as wildcard for any characters.
-/// Handles multiple wildcards (e.g., `*admin*`, `foo*bar*baz`).
-pub(crate) fn glob_match(pattern: &str, value: &str) -> bool {
-    let segments: Vec<&str> = pattern.split('*').collect();
-
-    if segments.len() == 1 {
-        return pattern == value;
-    }
-
-    let starts_with_star = pattern.starts_with('*');
-    let ends_with_star = pattern.ends_with('*');
-
-    let segments = segments.as_slice();
-
-    if !starts_with_star {
-        let first = segments[0];
-        if !value.starts_with(first) {
-            return false;
-        }
-        let rest = &value[first.len()..];
-        return match_middle_and_end(&segments[1..], rest, ends_with_star);
-    }
-
-    if !ends_with_star {
-        let last = segments[segments.len() - 1];
-        if !value.ends_with(last) {
-            return false;
-        }
-        let rest = &value[..value.len() - last.len()];
-        return match_middle(&segments[..segments.len() - 1], rest);
-    }
-
-    match_middle(segments, value)
-}
-
-fn match_middle(segments: &[&str], mut value: &str) -> bool {
-    for seg in segments.iter() {
-        if seg.is_empty() {
-            continue;
-        }
-        match value.find(seg) {
-            Some(pos) => {
-                value = &value[pos + seg.len()..];
-            }
-            None => return false,
-        }
-    }
-    true
-}
-
-fn match_middle_and_end(segments: &[&str], mut value: &str, ends_with_star: bool) -> bool {
-    if segments.is_empty() {
-        return true;
-    }
-
-    let count = if ends_with_star {
-        segments.len()
-    } else {
-        segments.len() - 1
-    };
-
-    for seg in &segments[..count] {
-        if seg.is_empty() {
-            continue;
-        }
-        match value.find(seg) {
-            Some(pos) => {
-                value = &value[pos + seg.len()..];
-            }
-            None => return false,
-        }
-    }
-
-    if !ends_with_star {
-        let last = segments[segments.len() - 1];
-        if last.is_empty() {
-            return true;
-        }
-        return value.ends_with(last);
-    }
-
-    true
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
 #[cfg(test)]
 mod tests {
+    use super::super::types::SubjectConfig;
     use super::*;
-    use crate::classifier::Source;
+    use crate::classifier::{Source, ToolClassification};
+    use std::collections::HashMap;
 
     fn alice() -> AuthIdentity {
         AuthIdentity::new("alice", vec!["admin".to_string()])
@@ -846,16 +478,6 @@ mod tests {
 
     fn anon() -> AuthIdentity {
         AuthIdentity::anonymous()
-    }
-
-    /// Shorthand: call is_tool_allowed and return just the bool.
-    fn is_tool_allowed_bool(
-        identity: &AuthIdentity,
-        tool_name: &str,
-        acl: &AclConfig,
-        ctx: Option<&ToolContext>,
-    ) -> bool {
-        is_tool_allowed(identity, tool_name, acl, ctx).allowed
     }
 
     fn make_classification(kind: Kind) -> ToolClassification {
@@ -873,20 +495,20 @@ mod tests {
 
     #[test]
     fn test_default_allow_no_rules() {
-        let acl = AclConfig::legacy(AclPolicy::Allow, vec![]);
-        assert!(is_tool_allowed_bool(&alice(), "any_tool", &acl, None));
-        assert!(is_tool_allowed_bool(&anon(), "any_tool", &acl, None));
+        let acl = super::super::AclConfig::legacy(AclPolicy::Allow, vec![]);
+        assert!(super::super::is_tool_allowed(&alice(), "any_tool", &acl, None).allowed);
+        assert!(super::super::is_tool_allowed(&anon(), "any_tool", &acl, None).allowed);
     }
 
     #[test]
     fn test_default_deny_no_rules() {
-        let acl = AclConfig::legacy(AclPolicy::Deny, vec![]);
-        assert!(!is_tool_allowed_bool(&alice(), "any_tool", &acl, None));
+        let acl = super::super::AclConfig::legacy(AclPolicy::Deny, vec![]);
+        assert!(!super::super::is_tool_allowed(&alice(), "any_tool", &acl, None).allowed);
     }
 
     #[test]
     fn test_deny_specific_subject() {
-        let acl = AclConfig::legacy(
+        let acl = super::super::AclConfig::legacy(
             AclPolicy::Allow,
             vec![AclRule {
                 subjects: vec!["bob".to_string()],
@@ -896,29 +518,18 @@ mod tests {
             }],
         );
 
-        assert!(!is_tool_allowed_bool(
-            &bob(),
-            "sentry__search_issues",
-            &acl,
-            None
-        ));
-        assert!(is_tool_allowed_bool(
-            &bob(),
-            "slack__send_message",
-            &acl,
-            None
-        ));
-        assert!(is_tool_allowed_bool(
-            &alice(),
-            "sentry__search_issues",
-            &acl,
-            None
-        ));
+        assert!(
+            !super::super::is_tool_allowed(&bob(), "sentry__search_issues", &acl, None).allowed
+        );
+        assert!(super::super::is_tool_allowed(&bob(), "slack__send_message", &acl, None).allowed);
+        assert!(
+            super::super::is_tool_allowed(&alice(), "sentry__search_issues", &acl, None).allowed
+        );
     }
 
     #[test]
     fn test_allow_specific_role() {
-        let acl = AclConfig::legacy(
+        let acl = super::super::AclConfig::legacy(
             AclPolicy::Deny,
             vec![AclRule {
                 subjects: vec![],
@@ -928,13 +539,13 @@ mod tests {
             }],
         );
 
-        assert!(is_tool_allowed_bool(&alice(), "anything", &acl, None));
-        assert!(!is_tool_allowed_bool(&bob(), "anything", &acl, None));
+        assert!(super::super::is_tool_allowed(&alice(), "anything", &acl, None).allowed);
+        assert!(!super::super::is_tool_allowed(&bob(), "anything", &acl, None).allowed);
     }
 
     #[test]
     fn test_first_match_wins() {
-        let acl = AclConfig::legacy(
+        let acl = super::super::AclConfig::legacy(
             AclPolicy::Allow,
             vec![
                 AclRule {
@@ -952,47 +563,15 @@ mod tests {
             ],
         );
 
-        assert!(is_tool_allowed_bool(
-            &bob(),
-            "sentry__search_issues",
-            &acl,
-            None
-        ));
-        assert!(!is_tool_allowed_bool(
-            &bob(),
-            "sentry__delete_project",
-            &acl,
-            None
-        ));
-    }
-
-    #[test]
-    fn test_glob_exact_match() {
-        assert!(glob_match("my_tool", "my_tool"));
-        assert!(!glob_match("my_tool", "other_tool"));
-    }
-
-    #[test]
-    fn test_glob_wildcard_all() {
-        assert!(glob_match("*", "anything"));
-    }
-
-    #[test]
-    fn test_glob_prefix_wildcard() {
-        assert!(glob_match("sentry__*", "sentry__search_issues"));
-        assert!(glob_match("sentry__*", "sentry__"));
-        assert!(!glob_match("sentry__*", "slack__send"));
-    }
-
-    #[test]
-    fn test_glob_suffix_wildcard() {
-        assert!(glob_match("*_issues", "search_issues"));
-        assert!(!glob_match("*_issues", "search_users"));
+        assert!(super::super::is_tool_allowed(&bob(), "sentry__search_issues", &acl, None).allowed);
+        assert!(
+            !super::super::is_tool_allowed(&bob(), "sentry__delete_project", &acl, None).allowed
+        );
     }
 
     #[test]
     fn test_wildcard_subject() {
-        let acl = AclConfig::legacy(
+        let acl = super::super::AclConfig::legacy(
             AclPolicy::Deny,
             vec![AclRule {
                 subjects: vec!["*".to_string()],
@@ -1002,20 +581,15 @@ mod tests {
             }],
         );
 
-        assert!(is_tool_allowed_bool(&alice(), "health__check", &acl, None));
-        assert!(is_tool_allowed_bool(&bob(), "health__check", &acl, None));
-        assert!(is_tool_allowed_bool(&anon(), "health__check", &acl, None));
-        assert!(!is_tool_allowed_bool(
-            &alice(),
-            "sentry__search",
-            &acl,
-            None
-        ));
+        assert!(super::super::is_tool_allowed(&alice(), "health__check", &acl, None).allowed);
+        assert!(super::super::is_tool_allowed(&bob(), "health__check", &acl, None).allowed);
+        assert!(super::super::is_tool_allowed(&anon(), "health__check", &acl, None).allowed);
+        assert!(!super::super::is_tool_allowed(&alice(), "sentry__search", &acl, None).allowed);
     }
 
     #[test]
     fn test_combined_subject_and_role() {
-        let acl = AclConfig::legacy(
+        let acl = super::super::AclConfig::legacy(
             AclPolicy::Allow,
             vec![AclRule {
                 subjects: vec!["bob".to_string()],
@@ -1025,86 +599,17 @@ mod tests {
             }],
         );
 
-        assert!(is_tool_allowed_bool(
-            &bob(),
-            "dangerous__delete",
-            &acl,
-            None
-        ));
+        assert!(super::super::is_tool_allowed(&bob(), "dangerous__delete", &acl, None).allowed);
 
         let bob_admin = AuthIdentity::new("bob", vec!["admin".to_string()]);
-        assert!(!is_tool_allowed_bool(
-            &bob_admin,
-            "dangerous__delete",
-            &acl,
-            None
-        ));
-    }
-
-    #[test]
-    fn test_glob_both_wildcards() {
-        assert!(glob_match("*admin*", "admin_tool"));
-        assert!(glob_match("*admin*", "my_admin_cmd"));
-        assert!(glob_match("*admin*", "admin"));
-        assert!(glob_match("*admin*", "administrator"));
-        assert!(!glob_match("*admin*", "adm"));
-        assert!(!glob_match("*admin*", "dmin"));
-    }
-
-    #[test]
-    fn test_glob_both_wildcards_with_literal() {
-        assert!(glob_match("*_admin_*", "sentry_admin_search"));
-        assert!(glob_match("*_admin_*", "x_admin_y"));
-        assert!(!glob_match("*_admin_*", "admin_ping"));
-        assert!(!glob_match("*_admin_*", "admin"));
-        assert!(!glob_match("*_admin_*", "admin_"));
-        assert!(!glob_match("*_admin_*", "_admin"));
-    }
-
-    #[test]
-    fn test_glob_prefix_and_suffix_no_bookend() {
-        assert!(glob_match("foo*bar", "foobar"));
-        assert!(glob_match("foo*bar", "fooXYZbar"));
-        assert!(!glob_match("foo*bar", "barfoo"));
-        assert!(!glob_match("foo*bar", "foo"));
-        assert!(!glob_match("foo*bar", "bar"));
-    }
-
-    #[test]
-    fn test_glob_multiple_wildcards() {
-        assert!(glob_match("a*b*c", "aXXXbYYYc"));
-        assert!(glob_match("a*b*c", "abc"));
-        assert!(!glob_match("a*b*c", "ac"));
-        assert!(!glob_match("a*b*c", "abcX"));
-        assert!(!glob_match("a*b*c", "Xabc"));
-        assert!(glob_match(
-            "sentry__*_admin__*",
-            "sentry__team_admin__delete"
-        ));
-        assert!(!glob_match("sentry__*_admin__*", "sentry__admin__list"));
-    }
-
-    #[test]
-    fn test_glob_empty_segments() {
-        assert!(glob_match("**", "anything"));
-        assert!(glob_match("**", ""));
-    }
-
-    #[test]
-    fn test_glob_regression_existing() {
-        assert!(glob_match("my_tool", "my_tool"));
-        assert!(!glob_match("my_tool", "other_tool"));
-        assert!(glob_match("*", "anything"));
-        assert!(glob_match("sentry__*", "sentry__search_issues"));
-        assert!(glob_match("sentry__*", "sentry__"));
-        assert!(!glob_match("sentry__*", "slack__send"));
-        assert!(glob_match("*_issues", "search_issues"));
-        assert!(!glob_match("*_issues", "search_users"));
+        assert!(
+            !super::super::is_tool_allowed(&bob_admin, "dangerous__delete", &acl, None).allowed
+        );
     }
 
     #[test]
     fn test_acl_with_contains_pattern() {
-        let acl = AclConfig::legacy(
+        let acl = super::super::AclConfig::legacy(
             AclPolicy::Allow,
             vec![AclRule {
                 subjects: vec!["bob".to_string()],
@@ -1114,171 +619,10 @@ mod tests {
             }],
         );
 
-        assert!(!is_tool_allowed_bool(&bob(), "admin_panel", &acl, None));
-        assert!(!is_tool_allowed_bool(
-            &bob(),
-            "user_admin_panel",
-            &acl,
-            None
-        ));
-        assert!(is_tool_allowed_bool(&bob(), "user_panel", &acl, None));
-        assert!(is_tool_allowed_bool(&alice(), "admin_panel", &acl, None));
-    }
-
-    #[test]
-    fn test_acl_config_deserialize() {
-        let json = r#"{
-            "default": "deny",
-            "rules": [
-                {
-                    "subjects": ["alice"],
-                    "tools": ["*"],
-                    "policy": "allow"
-                },
-                {
-                    "roles": ["viewer"],
-                    "tools": ["sentry__*"],
-                    "policy": "allow"
-                }
-            ]
-        }"#;
-        let acl: AclConfig = serde_json::from_str(json).unwrap();
-        match &acl {
-            AclConfig::Legacy(legacy) => {
-                assert_eq!(legacy.default, AclPolicy::Deny);
-                assert_eq!(legacy.rules.len(), 2);
-                assert_eq!(legacy.rules[0].subjects, vec!["alice"]);
-                assert!(legacy.rules[0].roles.is_empty());
-                assert_eq!(legacy.rules[1].roles, vec!["viewer"]);
-            }
-            AclConfig::RoleBased(_) => panic!("expected legacy schema"),
-        }
-    }
-
-    #[test]
-    fn test_acl_default_deserialize() {
-        let json = r#"{"rules": []}"#;
-        let acl: AclConfig = serde_json::from_str(json).unwrap();
-        match &acl {
-            AclConfig::Legacy(legacy) => assert_eq!(legacy.default, AclPolicy::Allow),
-            AclConfig::RoleBased(_) => panic!("expected legacy schema"),
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // Schema detection tests
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn test_deserialize_rbac_schema() {
-        let json = r#"{
-            "default": "deny",
-            "roles": {
-                "admin": [{ "server": "*", "access": "*" }]
-            },
-            "subjects": {
-                "alice": { "roles": ["admin"] }
-            }
-        }"#;
-        let acl: AclConfig = serde_json::from_str(json).unwrap();
-        match &acl {
-            AclConfig::RoleBased(rbac) => {
-                assert_eq!(rbac.default, AclPolicy::Deny);
-                assert!(rbac.roles.contains_key("admin"));
-                assert!(rbac.subjects.contains_key("alice"));
-            }
-            AclConfig::Legacy(_) => panic!("expected role-based schema"),
-        }
-    }
-
-    #[test]
-    fn test_deserialize_subjects_only_schema() {
-        let json = r#"{
-            "default": "deny",
-            "roles": { "dev": [{ "server": "*", "access": "read" }] },
-            "subjects": {
-                "bob": { "roles": ["dev"] }
-            }
-        }"#;
-        let acl: AclConfig = serde_json::from_str(json).unwrap();
-        assert!(matches!(acl, AclConfig::RoleBased(_)));
-    }
-
-    #[test]
-    fn test_deserialize_both_schemas_error() {
-        let json = r#"{
-            "rules": [{"subjects": ["x"], "tools": ["*"], "policy": "allow"}],
-            "roles": { "admin": [{ "server": "*", "access": "*" }] }
-        }"#;
-        let result: Result<AclConfig, _> = serde_json::from_str(json);
-        assert!(result.is_err());
-        let err = result.unwrap_err().to_string();
-        assert!(err.contains("cannot have both"));
-    }
-
-    #[test]
-    fn test_deserialize_empty_acl() {
-        let json = r#"{}"#;
-        let acl: AclConfig = serde_json::from_str(json).unwrap();
-        match &acl {
-            AclConfig::Legacy(legacy) => {
-                assert_eq!(legacy.default, AclPolicy::Allow);
-                assert!(legacy.rules.is_empty());
-            }
-            AclConfig::RoleBased(_) => panic!("expected legacy for empty config"),
-        }
-    }
-
-    #[test]
-    fn test_deserialize_rbac_strict_classification() {
-        let json = r#"{
-            "default": "deny",
-            "strictClassification": true,
-            "roles": {}
-        }"#;
-        let acl: AclConfig = serde_json::from_str(json).unwrap();
-        match &acl {
-            AclConfig::RoleBased(rbac) => assert!(rbac.strict_classification),
-            AclConfig::Legacy(_) => panic!("expected role-based schema"),
-        }
-    }
-
-    #[test]
-    fn test_deserialize_rbac_full_example() {
-        let json = r#"{
-            "default": "deny",
-            "strictClassification": false,
-            "roles": {
-                "admin":    [{ "server": "*", "access": "*" }],
-                "dev": [
-                    { "server": ["github", "grafana"], "access": "read" },
-                    { "server": "github", "access": "write", "tools": ["gh_pr", "gh_issue"] }
-                ],
-                "readonly": [{ "server": "*", "access": "read" }]
-            },
-            "subjects": {
-                "alice":   { "roles": ["admin"] },
-                "bob":     { "roles": ["dev"] },
-                "charlie": {
-                    "roles": ["readonly"],
-                    "extra": [{ "server": "sentry", "access": "read", "resources": ["issue://*"] }]
-                }
-            }
-        }"#;
-        let acl: AclConfig = serde_json::from_str(json).unwrap();
-        match &acl {
-            AclConfig::RoleBased(rbac) => {
-                assert_eq!(rbac.roles.len(), 3);
-                assert_eq!(rbac.subjects.len(), 3);
-                assert!(!rbac.strict_classification);
-                let dev_grants = &rbac.roles["dev"];
-                assert_eq!(dev_grants.len(), 2);
-                let charlie = &rbac.subjects["charlie"];
-                assert_eq!(charlie.extra.len(), 1);
-                assert_eq!(charlie.extra[0].resources, vec!["issue://*"]);
-            }
-            AclConfig::Legacy(_) => panic!("expected role-based schema"),
-        }
+        assert!(!super::super::is_tool_allowed(&bob(), "admin_panel", &acl, None).allowed);
+        assert!(!super::super::is_tool_allowed(&bob(), "user_admin_panel", &acl, None).allowed);
+        assert!(super::super::is_tool_allowed(&bob(), "user_panel", &acl, None).allowed);
+        assert!(super::super::is_tool_allowed(&alice(), "admin_panel", &acl, None).allowed);
     }
 
     // -----------------------------------------------------------------------
@@ -2040,8 +1384,6 @@ mod tests {
 
     #[test]
     fn test_deny_read_does_not_block_write() {
-        // A deny grant with access=read should only block read tools,
-        // not write tools on the same server/tool pattern.
         let read_cls = make_classification(Kind::Read);
         let write_cls = make_classification(Kind::Write);
         let acl = RoleBasedAclConfig {
@@ -2080,7 +1422,7 @@ mod tests {
         };
         assert!(!is_tool_allowed_rbac(&identity, &ctx, &acl).allowed);
 
-        // Write is NOT denied — the deny grant only covers read
+        // Write is NOT denied -- the deny grant only covers read
         let ctx = ToolContext {
             server_alias: "github",
             tool_name: "gh_secret_tool",
@@ -2129,7 +1471,7 @@ mod tests {
         };
         assert!(!is_tool_allowed_rbac(&identity, &ctx, &acl).allowed);
 
-        // Read is NOT denied — the deny grant only covers write
+        // Read is NOT denied -- the deny grant only covers write
         let ctx = ToolContext {
             server_alias: "github",
             tool_name: "gh_repo_delete",
@@ -2138,78 +1480,13 @@ mod tests {
         assert!(is_tool_allowed_rbac(&identity, &ctx, &acl).allowed);
     }
 
-    #[test]
-    fn test_deserialize_roles_wrong_type_errors() {
-        // roles as array (not object) should trigger new schema detection
-        // and then fail deserialization, not silently fall to legacy.
-        let json = r#"{
-            "default": "deny",
-            "roles": ["admin"]
-        }"#;
-        let result: Result<AclConfig, _> = serde_json::from_str(json);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_deserialize_subjects_wrong_type_errors() {
-        let json = r#"{
-            "default": "deny",
-            "subjects": "alice"
-        }"#;
-        let result: Result<AclConfig, _> = serde_json::from_str(json);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_deserialize_unknown_role_reference_errors() {
-        // Subject references a role that doesn't exist in the roles map.
-        let json = r#"{
-            "default": "deny",
-            "roles": {
-                "admin": [{ "server": "*", "access": "*" }]
-            },
-            "subjects": {
-                "alice": { "roles": ["nonexistent"] }
-            }
-        }"#;
-        let result: Result<AclConfig, _> = serde_json::from_str(json);
-        assert!(result.is_err());
-        let err = result.unwrap_err().to_string();
-        assert!(err.contains("unknown role"));
-    }
-
     // -----------------------------------------------------------------------
     // Decision provenance tests
     // -----------------------------------------------------------------------
 
     #[test]
-    fn test_matched_rule_display() {
-        assert_eq!(MatchedRule::Legacy(0).to_string(), "legacy[0]");
-        assert_eq!(MatchedRule::Legacy(3).to_string(), "legacy[3]");
-        assert_eq!(MatchedRule::LegacyDefault.to_string(), "legacy:default");
-        assert_eq!(
-            MatchedRule::RoleGrant {
-                role: "dev".to_string(),
-                index: 1
-            }
-            .to_string(),
-            "dev[1]"
-        );
-        assert_eq!(
-            MatchedRule::SubjectExtra {
-                subject: "alice".to_string(),
-                index: 0
-            }
-            .to_string(),
-            "alice.extra[0]"
-        );
-        assert_eq!(MatchedRule::RbacDefault.to_string(), "default");
-        assert_eq!(MatchedRule::NoAcl.to_string(), "no-acl");
-    }
-
-    #[test]
     fn test_decision_legacy_first_rule_match() {
-        let acl = AclConfig::legacy(
+        let acl = super::super::AclConfig::legacy(
             AclPolicy::Allow,
             vec![AclRule {
                 subjects: vec!["bob".to_string()],
@@ -2218,14 +1495,14 @@ mod tests {
                 policy: AclPolicy::Deny,
             }],
         );
-        let d = is_tool_allowed(&bob(), "sentry__search_issues", &acl, None);
+        let d = super::super::is_tool_allowed(&bob(), "sentry__search_issues", &acl, None);
         assert!(!d.allowed);
         assert_eq!(d.matched_rule, MatchedRule::Legacy(0));
     }
 
     #[test]
     fn test_decision_legacy_second_rule_match() {
-        let acl = AclConfig::legacy(
+        let acl = super::super::AclConfig::legacy(
             AclPolicy::Allow,
             vec![
                 AclRule {
@@ -2242,14 +1519,14 @@ mod tests {
                 },
             ],
         );
-        let d = is_tool_allowed(&bob(), "sentry__delete", &acl, None);
+        let d = super::super::is_tool_allowed(&bob(), "sentry__delete", &acl, None);
         assert!(!d.allowed);
         assert_eq!(d.matched_rule, MatchedRule::Legacy(1));
     }
 
     #[test]
     fn test_decision_legacy_default_fallback() {
-        let acl = AclConfig::legacy(
+        let acl = super::super::AclConfig::legacy(
             AclPolicy::Deny,
             vec![AclRule {
                 subjects: vec!["alice".to_string()],
@@ -2258,7 +1535,7 @@ mod tests {
                 policy: AclPolicy::Allow,
             }],
         );
-        let d = is_tool_allowed(&bob(), "anything", &acl, None);
+        let d = super::super::is_tool_allowed(&bob(), "anything", &acl, None);
         assert!(!d.allowed);
         assert_eq!(d.matched_rule, MatchedRule::LegacyDefault);
     }
@@ -2451,13 +1728,6 @@ mod tests {
         assert!((d.classification_confidence.unwrap() - 0.72).abs() < f32::EPSILON);
     }
 
-    #[test]
-    fn test_decision_access_level_as_str() {
-        assert_eq!(AccessLevel::Read.as_str(), "read");
-        assert_eq!(AccessLevel::Write.as_str(), "write");
-        assert_eq!(AccessLevel::All.as_str(), "*");
-    }
-
     // -----------------------------------------------------------------------
     // Resource ACL tests
     // -----------------------------------------------------------------------
@@ -2535,7 +1805,7 @@ mod tests {
         assert!(!matches_prompt_grant(&grant, "translate"));
     }
 
-    fn rbac_with_resource_grants() -> AclConfig {
+    fn rbac_with_resource_grants() -> super::super::AclConfig {
         let json = serde_json::json!({
             "default": "deny",
             "roles": {
@@ -2629,7 +1899,7 @@ mod tests {
                 "bob": { "roles": ["restricted"] }
             }
         });
-        let acl: AclConfig = serde_json::from_value(json).unwrap();
+        let acl: super::super::AclConfig = serde_json::from_value(json).unwrap();
         let ctx = ResourceContext {
             server_alias: "sentry",
             resource_uri: "secret://key",
@@ -2640,7 +1910,7 @@ mod tests {
 
     #[test]
     fn test_resource_legacy_deny_default() {
-        let acl = AclConfig::legacy(AclPolicy::Deny, vec![]);
+        let acl = super::super::AclConfig::legacy(AclPolicy::Deny, vec![]);
         let ctx = ResourceContext {
             server_alias: "sentry",
             resource_uri: "issue://123",
@@ -2652,7 +1922,7 @@ mod tests {
 
     #[test]
     fn test_resource_legacy_allow_default() {
-        let acl = AclConfig::legacy(AclPolicy::Allow, vec![]);
+        let acl = super::super::AclConfig::legacy(AclPolicy::Allow, vec![]);
         let ctx = ResourceContext {
             server_alias: "sentry",
             resource_uri: "issue://123",
@@ -2690,7 +1960,7 @@ mod tests {
                 "bob": { "roles": ["writer"] }
             }
         });
-        let acl: AclConfig = serde_json::from_value(json).unwrap();
+        let acl: super::super::AclConfig = serde_json::from_value(json).unwrap();
         let ctx = ResourceContext {
             server_alias: "sentry",
             resource_uri: "issue://123",
@@ -2704,7 +1974,7 @@ mod tests {
     // Prompt ACL tests
     // -----------------------------------------------------------------------
 
-    fn rbac_with_prompt_grants() -> AclConfig {
+    fn rbac_with_prompt_grants() -> super::super::AclConfig {
         let json = serde_json::json!({
             "default": "deny",
             "roles": {
@@ -2789,7 +2059,7 @@ mod tests {
                 "bob": { "roles": ["restricted"] }
             }
         });
-        let acl: AclConfig = serde_json::from_value(json).unwrap();
+        let acl: super::super::AclConfig = serde_json::from_value(json).unwrap();
         let ctx = PromptContext {
             server_alias: "ai",
             prompt_name: "dangerous_prompt",
@@ -2800,7 +2070,7 @@ mod tests {
 
     #[test]
     fn test_prompt_legacy_deny_default() {
-        let acl = AclConfig::legacy(AclPolicy::Deny, vec![]);
+        let acl = super::super::AclConfig::legacy(AclPolicy::Deny, vec![]);
         let ctx = PromptContext {
             server_alias: "ai",
             prompt_name: "summarize",
@@ -2811,7 +2081,7 @@ mod tests {
 
     #[test]
     fn test_prompt_legacy_allow_default() {
-        let acl = AclConfig::legacy(AclPolicy::Allow, vec![]);
+        let acl = super::super::AclConfig::legacy(AclPolicy::Allow, vec![]);
         let ctx = PromptContext {
             server_alias: "ai",
             prompt_name: "summarize",
